+++ R/zzz.R
@@ -3,7 +3,8 @@
 .onLoad <- function(libname, pkgname) {
   opts <- options()
   our_defaults <- list(
-    rquery.verbose = FALSE
+    rquery.verbose = FALSE,
+    rquery.rquery_db_executor = NULL
   )
   defs <- setdiff(names(our_defaults),
                   names(opts))